aggregate DelegateAttribute to Attributes module
